requirements.txt
django~=3.1.3
gunicorn
django-heroku
djangorestframework~=3.12.2